% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPermPalindrome.R
\name{isPermPalindrome}
\alias{isPermPalindrome}
\title{isPermPalindrome Function}
\usage{
isPermPalindrome(input)
}
\arguments{
\item{input:}{a string input as a dataframe}
}
\value{
isPermPalindrome(input) --> boolean T/F
}
\description{
Function check whether if the input is a palindromic word or not
}
