% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingNOs.R
\name{MissingNOs}
\alias{MissingNOs}
\title{MissingNOs Function}
\usage{
MissingNOs(missing_Numbers)
}
\arguments{
\item{missing_Numbers:}{A vector input that contains numbers in range 1-100 with missings}
}
\value{
missingNOs([1,2, 3, 100) => missingNOs(missing_Numbers)
}
\description{
Function creates a vector to compare with input. If there are any missing values, it prints the missing ones
}
