% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productOfTheOthers.R
\name{productOfTheOthers}
\alias{productOfTheOthers}
\title{productOfTheOthers Function}
\usage{
productOfTheOthers(num_array)
}
\arguments{
\item{num_array:}{an input vector}
}
\value{
productOfTheOthers(num_array)
}
\description{
Function creates a vector that have same length with the first vector and return num_array which is the multiplication of the others.
}
